#17 feat: add dialog if user isn't logged in
This commit adds a dialog box that prompts users to log in before adding a beverage. The dialog appears when the user clicks on the "add" button and is not logged in. The dialog includes an option to log in or cancel, and redirects the user to the login page upon clicking "log in".