#8 feat: add toggle button to hide password
This commit adds a toggle button to the login, register and reset password pages that allows users to hide or show their passwords. It also updates the i18n files with new translations. Finally it updates the validation rules for the input fields on these pages to use these translations instead of hardcoded strings.